Move Connector implementations to api/connector package
