+++ api/connector/simpleconnector.go
@@ -1,18 +1,13 @@
 // Copyright 2022 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package api
+package connector
 
 import (
+	"github.com/juju/juju/api"
 	"github.com/juju/names/v4"
 	"gopkg.in/macaroon.v2"
 )
-
-// A Connector is able to provide a Connection.  This connection can be used to
-// make API calls via the various packages in github.com/juju/juju/api.
-type Connector interface {
-	Connect(...DialOption) (Connection, error)
-}
 
 // SimpleConnectorConfig aims to provide the same API surface as pilot juju for
 // obtaining an api connection.
@@ -35,17 +30,17 @@ type SimpleConnectorConfig struct {
 
 // A SimpleConnector can provide connections from a simple set of options.
 type SimpleConnector struct {
-	info            Info
-	defaultDialOpts DialOpts
+	info            api.Info
+	defaultDialOpts api.DialOpts
 }
 
 var _ Connector = (*SimpleConnector)(nil)
 
-// NewSimpleConnector returns an instance of *SimpleConnector configured to
+// NewSimple returns an instance of *SimpleConnector configured to
 // connect according to the specified options.  If some options are invalid an
 // error is returned.
-func NewSimpleConnector(opts SimpleConnectorConfig, dialOptions ...DialOption) (*SimpleConnector, error) {
-	info := Info{
+func NewSimple(opts SimpleConnectorConfig, dialOptions ...api.DialOption) (*SimpleConnector, error) {
+	info := api.Info{
 		Addrs:    []string{opts.ControllerAddress},
 		CACert:   opts.CACert,
 		ModelTag: names.NewModelTag(opts.ModelUUID),
@@ -59,7 +54,7 @@ func NewSimpleConnector(opts SimpleConnectorConfig, dialOptions ...DialOption) (
 	}
 	conn := &SimpleConnector{
 		info:            info,
-		defaultDialOpts: DefaultDialOpts(),
+		defaultDialOpts: api.DefaultDialOpts(),
 	}
 	for _, f := range dialOptions {
 		f(&conn.defaultDialOpts)
@@ -68,10 +63,10 @@ func NewSimpleConnector(opts SimpleConnectorConfig, dialOptions ...DialOption) (
 }
 
 // Connect returns a Connection according to c's configuration.
-func (c *SimpleConnector) Connect(dialOptions ...DialOption) (Connection, error) {
+func (c *SimpleConnector) Connect(dialOptions ...api.DialOption) (api.Connection, error) {
 	opts := c.defaultDialOpts
 	for _, f := range dialOptions {
 		f(&opts)
 	}
-	return Open(&c.info, c.defaultDialOpts)
+	return api.Open(&c.info, c.defaultDialOpts)
 }
